// src/app/api/report/path.ts
export enum PATHS {
	BILLING = '/report/billing',
	USER = '/report/user',
	PATIENT = '/report/patient',
	EXAMINATION = '/report/examination',
	ANALYSIS = '/report/analysis',
	DIAGNOSIS = '/report/diagnosis',
	DOCTOR = '/report/doctor',
	ROOM = '/report/room',
	BLOOD_PRESSURE = '/report/blood-pressure',
}